Refactor stats module into submodules
