+++ src/Tools/Stats/__init__.py
@@ -1,9 +1,14 @@
 # src/Tools/Stats/__init__.py
 
-from .stats           import StatsAnalysisWindow
-from .stats_export    import export_significance_results_to_excel
+from .stats import StatsAnalysisWindow
+from .stats_export import export_significance_results_to_excel
+from . import stats_file_scanner, stats_ui, stats_runners, stats_helpers
 
 __all__ = [
     "StatsAnalysisWindow",
     "export_significance_results_to_excel",
+    "stats_file_scanner",
+    "stats_ui",
+    "stats_runners",
+    "stats_helpers",
 ]
